requirements.txt
playwright
pytest-playwright
pytest
pytest-bdd
pytest-gherkin
allure-pytest
allure-python-commons
pytest-selenium
requests
python-dotenv
qase-pytest
qase-api-client
